test: move the example components inside test/vitest
