+++ packages/unit-vitest/src/templates/base/test/vitest/___tests__/ExampleComponent.vitest.test.ts
@@ -1,7 +1,7 @@
 import { installQuasar } from '@quasar/quasar-app-extension-testing-unit-vitest';
 import { mount } from '@vue/test-utils';
 import { describe, expect, it } from 'vitest';
-import ExampleComponent from '../ExampleComponent.vue';
+import ExampleComponent from './demo/ExampleComponent.vue';
 
 installQuasar();
