+++ packages/unit-vitest/src/templates/base/test/vitest/___tests__/demo/ExampleComponent.vue
