+++ test-project-vite/package.json
