+++ test-project-vite/quasar.extensions.json
@@ -1,15 +1,28 @@
 {
   "@quasar/testing": {
-    "harnesses": ["unit-jest@alpha", "e2e-cypress"]
+    "harnesses": [
+      "unit-jest@alpha",
+      "e2e-cypress"
+    ]
   },
   "@quasar/testing-unit-jest": {
     "babel": "babelrc",
-    "options": ["scripts", "typescript", "majestic"]
+    "options": [
+      "scripts",
+      "typescript",
+      "majestic"
+    ]
   },
   "@quasar/testing-e2e-cypress": {
-    "options": ["scripts", "typescript", "code-coverage"]
+    "options": [
+      "scripts",
+      "typescript",
+      "code-coverage"
+    ]
   },
   "@quasar/testing-unit-vitest": {
-    "options": ["ui"]
+    "options": [
+      "ui"
+    ]
   }
 }
