+++ test-project-vite/test/vitest/__tests__/NotifyComponent.vitest.test.ts
@@ -2,7 +2,7 @@ import { installQuasar } from '@quasar/quasar-app-extension-testing-unit-vitest'
 import { mount } from '@vue/test-utils';
 import { Notify } from 'quasar';
 import { describe, expect, it, vi } from 'vitest';
-import NotifyComponent from '../NotifyComponent.vue';
+import NotifyComponent from './demo/NotifyComponent.vue';
 
 installQuasar({ plugins: { Notify } });
