+++ test-project-vite/test/vitest/__tests__/demo/ExampleComponent.vue
